# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Issue Tracker - API
  version: 1.0.0
  description: OpenAPI-like summary for the local AI Issue Tracker API routes
servers:
  - url: http://localhost:3000
paths:
...
  /api/ai/resolutions/{id}/rate:
    post:
      summary: Rate an AI resolution
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                feedback:
                  type: string
      responses:
        '200':
          description: Rating accepted
...
  /api/ai/retrieve:
    post:
      summary: Retrieve relevant KB entries (local RAG)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                contentType:
                  type: string
                limit:
                  type: integer
      responses:
        '200':
          description: Relevant entries and related issues
...
  /api/ai/generate-sop:
    post:
      summary: Generate SOP from resolution (Ollama)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                issueId:
                  type: string
      responses:
        '200':
          description: Created SOP
...